ref(ddm): Rename the files for metrics metadata
