+++ src/sentry/api/endpoints/organization_ddm.py
@@ -10,18 +10,18 @@ from sentry.api.api_publish_status import ApiPublishStatus
 from sentry.api.base import region_silo_endpoint
 from sentry.api.bases.organization import OrganizationEndpoint
 from sentry.api.serializers import serialize
-from sentry.api.serializers.models.code_locations import CodeLocationsSerializer
-from sentry.api.serializers.models.correlations import CorrelationsSerializer
+from sentry.api.serializers.models.metrics_code_locations import MetricCodeLocationsSerializer
+from sentry.api.serializers.models.metrics_correlations import MetricCorrelationsSerializer
 from sentry.api.utils import get_date_range_from_params
 from sentry.exceptions import InvalidParams
 from sentry.models.organization import Organization
 from sentry.models.project import Project
 from sentry.sentry_metrics.querying.errors import LatestReleaseNotFoundError
-from sentry.sentry_metrics.querying.metadata.code_locations import (
+from sentry.sentry_metrics.querying.metadata.metrics_code_locations import (
     MetricCodeLocations,
     get_metric_code_locations,
 )
-from sentry.sentry_metrics.querying.metadata.correlations import (
+from sentry.sentry_metrics.querying.metadata.metrics_correlations import (
     MetricCorrelations,
     get_metric_correlations,
 )
@@ -34,8 +34,8 @@ class MetricMetaType(Enum):
 
 
 METRIC_META_TYPE_SERIALIZER = {
-    MetricMetaType.CODE_LOCATIONS.value: CodeLocationsSerializer(),
-    MetricMetaType.CORRELATIONS.value: CorrelationsSerializer(),
+    MetricMetaType.CODE_LOCATIONS.value: MetricCodeLocationsSerializer(),
+    MetricMetaType.CORRELATIONS.value: MetricCorrelationsSerializer(),
 }
